(feature/handle-parser) Trying to add the ability to use both ALTO and Page [WIP]
